+++ scripts/systemd_user_cleanup.sh
@@ -18,6 +18,12 @@ for f in ${FILES}
 do
   if [[ -e /etc/systemd/user/${f} ]]
   then
+  	  su ${USER_NAME} -c "XDG_RUNTIME_DIR=\"/run/user/$(id -u ${USER_NAME})\" DBUS_SESSION_BUS_ADDRESS=\"unix:path=${XDG_RUNTIME_DIR}/bus\" systemctl --user disable ${f}"
+  	  if [[ ! $? -eq 0 ]]
+  	  then
+  	  	  echo -e "\nFailed whilst disabling systemd units."
+  	  	  exit 1
+  	  fi
       rm -rf /etc/systemd/user/${f}
   fi
 done
